+++ ospd/command/command.py
@@ -412,22 +412,25 @@ class GetScans(BaseCommand):
         details = xml.get('details')
         pop_res = xml.get('pop_results')
         max_res = xml.get('max_results')
+        progress = xml.get('progress')
 
         if details and details == '0':
             details = False
         else:
             details = True
-            if pop_res and pop_res == '1':
-                pop_res = True
-            else:
-                pop_res = False
+            pop_res = pop_res and pop_res == '1'
+
             if max_res:
                 max_res = int(max_res)
+
+        progress = progress and progress == '1'
 
         responses = []
         if scan_id and scan_id in self._daemon.scan_collection.ids_iterator():
             self._daemon.check_scan_process(scan_id)
-            scan = self._daemon.get_scan_xml(scan_id, details, pop_res, max_res)
+            scan = self._daemon.get_scan_xml(
+                scan_id, details, pop_res, max_res, progress
+            )
             responses.append(scan)
         elif scan_id:
             text = "Failed to find scan '{0}'".format(scan_id)
@@ -436,7 +439,7 @@ class GetScans(BaseCommand):
             for scan_id in self._daemon.scan_collection.ids_iterator():
                 self._daemon.check_scan_process(scan_id)
                 scan = self._daemon.get_scan_xml(
-                    scan_id, details, pop_res, max_res
+                    scan_id, details, pop_res, max_res, progress
                 )
                 responses.append(scan)